Squashed 'CPF/' changes from c52f8633..e0e05c95
e0e05c95 Fixed test for interval timer service, it ignores cancels correctly now. Changed how event emitter works a little.  It is now generic and there is no longer a deferred variation, behavior is controlled by the event sink which is installed.
978c5c0b Fixed a silly bug in the scheduler.  Forgot to adjust the minimal sample time to be proportional to the sample rate.
18b6781a Big build update. * Added static/dynamic variations of examples. * Most things use the new plugin install system. * Can now disable most static builds, great speed up of iteration times.  (Turn off CPF_BUILD_STATIC_VARIANTS.)
916321cc Moved the flocking app to use the new plugin installation system. Not sure if I'll make this the primary solution or just an option, though I might add direct support of it into the application class since that would solve some other problems.
80c9a025 Almost working attempt at the simplified registration system.
7b3db801 Some work looking at different clean up options.
9c4a8e55 Fleshed out some of the timer items. Updated the event driven example to use the timer system with the intent of starting on the real work for futures.
34a9160f Finished off the timer service.  Seems workable.
62c814ba Starting work on a timer service for concurrency uses. This should provide the starting async service which can be used to flesh out the async futures system.
2d345d1a Some work setting up the balance controllers and some changes to the thread controller itself.
5031fb98 Cleaned up some outdated items. Made the dev gui collapse panel sizes as possible.
cda138eb Balancing seems to work working, only tested on the thread ripper at this time though.
b5614e63 Close but not quite there.
fcac94f1 Almost have a working utilization solution.
640c54de Changes to the scheduler opcode wait, it no longer tries to use a backoff internally. Add self balancing to the flocking app as a basic test.
526cd9ae Some cleanup of the time interfaces. Added a utilization test to the scheduler.  It is causing too much overhead though and will need to be changed.
dacc89d6 quick cleanup of the reactor system.
79f073ce A bit of this, a bit of that.
66f24f73 Ripped out experiments, they were invalid.
5ca086db Messing around with options for the unification of the event loop solution.
830eac98 Some outlining of an approach to rust like futures.
120a09b4 Added a note to outline some thoughts on reworking futures to be more appropriate to the realtime nature of the scheduler.
0ebc124a Changed build flags to test some things. Fixed an issue with reverse iterators in utf8 strings. Added some more functions to the gui system for experiments being done.
3052b217 Basic outline and initial thinking.
18fb04d5 Should compile without D3D12 on out of date OS's, such as Windows 7.. :) Adding an example for event driven applications, this will form the basis for the work needed to add proper promise/future driven IO and related items.
76159c4a Renamed and namespaced entity manager.
2c59afae Some class factory cleanup and expansion. Needed for some intended dev tools.
34dd4775 Added iterator and basic tests. The component add/remove test will be a good KPI in the future.
b3ce89b7 Maybe the final tweek.
6aad1e03 Another tweek to build setup.
804f0c1f Minor build tweek for build machine without java.
cf9ef5cb I think 'yay' is in order.  Highly compact component storage seems to be working. Will still need to add a lazy optimizer to the component order. Also still need to recreate the component groups API's, though this time it will be 'iterators' I think.
32f6d527 Some more work on component containers.  The sparse stuff is being a pain and needs cleanup.
2f8cc055 Killed off some warnings. Adding further tests for the packed component replacements.
a13f61ff Reworking component storage and related.
5edfb71a Just a note about the command list problem and the probable correct solution for later.
48db1fe6 A hacky (and slowish) solution to the bad command lists in the UI.
8f8b45b9 Adding custom draw calls.
c747b25a Commenting some data block api's. Cleaning up some collision mesh builder.
096baedb Merge branch 'develop'
dd10b296 Initial resource cache and unit test.
43ff2f07 Some documentation updates. A bit of work on the cache api as it had a name conflict.
8b05406e Create README.md
2a2460ce Initial resource cache setup. Seems like it will work reasonably.
1527fcac Corrected the transform component data. Added project for the new resource management system. Further fiddling with how the cache system will work.
8bcfbe79 Early collision geometry building. Starting to look at the new resource system and various ideas behind it.
40651904 Physics example using the new CPFS formatted built meshes.
a776f719 Added write order to the DataBlocks.  Currently setup to be written in the same order as they are added. Added memory stream and initial CPFS container format.
871774cc Added a temporary header and method to test that the written datablocks function.  It's working.

git-subtree-dir: CPF
git-subtree-split: e0e05c956446c2b92f78d94d702449a210dd1e61

+++ Plugins/Drivers/PlatformGraphicsD3D12/Source/D3D12/Image.cpp
@@ -60,7 +60,7 @@ Image::Image(Device* device, Graphics::HeapType heap, const Graphics::ClearValue
 			else
 			{
 				clearValue.DepthStencil.Depth = cv->mData.mDepthStencil.mDepth;
-				clearValue.DepthStencil.Stencil = cv->mData.mDepthStencil.mStencil;
+				clearValue.DepthStencil.Stencil = UINT8(cv->mData.mDepthStencil.mStencil);
 			}
 		}
 		device->GetD3DDevice()->CreateCommittedResource(
